// application/frontend/src/app/shared/models/capacity-quantity.ts
/**
 * @license
 * Copyright 2022 Google LLC
 *
 * Use of this source code is governed by an MIT-style
 * license that can be found in the LICENSE file or at
 * https://opensource.org/licenses/MIT.
 */

export interface CapacityQuantityFormValue {
  type?: string;
  value?: number;
}

export interface CapacityQuantitySplitFormValue {
  type: string;
  value: number;
  unit?: string;
}